pyproject.toml
[project]
name = "job-application-automation"
version = "0.1.0"
description = "Automated pipeline that monitors, filters, and prepares job applications, delivered via Discord."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Jordan Shuhandler" }]

# Core runtime dependencies needed by every phase.
# Feature-specific dependencies live under [project.optional-dependencies].
dependencies = [
    "pydantic[email]>=2.7",
    "pydantic-settings>=2.4",
    "PyYAML>=6.0",
    "SQLAlchemy>=2.0",
    "alembic>=1.13",
    "typer>=0.12",
    "rich>=13.7",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
# Phase 1 — scraping / HTTP
scraping = [
    "httpx>=0.27",
    "parsel>=1.9",
    "selectolax>=0.3.21",
    "playwright>=1.45",
    "playwright-stealth>=2.0",
]

# Phase 2 — NLP / filtering
nlp = [
    "spacy>=3.7",
    "rapidfuzz>=3.9",
]

# Phase 1/5 — task queue
celery = [
    "celery[redis]>=5.4",
    "redis>=5.0",
]

# Phase 5 — Discord bot
discord = [
    "discord.py>=2.4",
]

# Phase 4 — PDF rendering for cover letters
pdf = [
    "weasyprint>=62.0",
    "jinja2>=3.1",
]

# Phase 4 — LLM integration (deferred; see D9)
llm = [
    "litellm>=1.40",
]

# Phase 6 — Gmail API
email = [
    "google-api-python-client>=2.130",
    "google-auth>=2.30",
    "google-auth-oauthlib>=1.2",
]

# Postgres driver (used in prod; SQLite by default in dev)
postgres = [
    "psycopg[binary]>=3.2",
]

# Convenience meta-extra: everything needed to run the full pipeline.
all = [
    "job-application-automation[scraping,nlp,celery,discord,pdf,llm,email,postgres]",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pre-commit>=3.8",
    "types-PyYAML>=6.0",
    "types-python-dateutil>=2.9",
]

[project.scripts]
jaa = "src.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ---------------------------------------------------------------------------
# uv
# ---------------------------------------------------------------------------
[tool.uv]
package = true

# ---------------------------------------------------------------------------
# Ruff — lint + format
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "B",             # flake8-bugbear
    "UP",            # pyupgrade
    "N",             # pep8-naming
    "SIM",           # flake8-simplify
    "RUF",           # ruff-specific
]
ignore = [
    "E501",          # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["N802", "N803", "N806"]
"src/db/migrations/**" = ["N999", "E402"]
# Typer uses function-call defaults (typer.Option / typer.Argument) as its
# canonical API. Silence B008 in the CLI surface only.
"src/cli/**" = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["src/db/migrations/versions"]

[[tool.mypy.overrides]]
module = [
    "parsel.*",
    "playwright_stealth.*",
    "selectolax.*",
    "celery.*",
    "weasyprint.*",
    "googleapiclient.*",
    "google_auth_oauthlib.*",
]
ignore_missing_imports = true

# Celery decorators are untyped; that's out of our control.
[[tool.mypy.overrides]]
module = ["src.tasks.*"]
disallow_untyped_decorators = false

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
asyncio_mode = "auto"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:passlib.*",
    # httpx/asyncio often leak sockets + event loops at test teardown
    # when asyncio.run() is used inside tests.
    "ignore::ResourceWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["src/db/migrations/*"]
